frontend: replace broccoli by esbuild
