+++ package.json
@@ -1,18 +1,25 @@
 {
+  "name": "congame",
+  "engines": {
+    "node": ">=16",
+    "npm": ">=7"
+  },
   "devDependencies": {
-    "broccoli": "^3.0.0",
-    "broccoli-cli": "^1.0.0",
-    "broccoli-concat": "^3.7.3",
-    "broccoli-funnel": "^2.0.2",
-    "broccoli-merge-trees": "^3.0.2",
-    "broccoli-sass-source-maps": "^4.0.0",
-    "broccoli-uglify-sourcemap": "^3.1.1",
-    "eslint": "^5.15.0",
-    "node-sass": "^7.0.1",
-    "prettier": "^1.16.4"
+    "esbuild": "^0.20.2",
+    "esbuild-plugin-copy": "^2.1.1",
+    "esbuild-sass-plugin": "^3.2.0",
+    "eslint": "^8.10.0",
+    "prettier": "^3.0.0",
+    "typescript": "^5.4.2"
   },
   "scripts": {
-    "build": "broccoli b --output-path static",
-    "watch": "broccoli b --output-path static --watch"
+    "build": "NODE_ENV=production node build.mjs",
+    "lint": "tsc -noEmit resources/js/app.ts",
+    "watch": "node build.mjs"
+  },
+  "dependencies": {
+    "@lit/task": "^1.0.0",
+    "lit": "^3.1.2",
+    "unpoly": "^3.7.3"
   }
 }
