+++ resources/js/sentry/sentry.ts
@@ -1,0 +1,1 @@
+export default (<any>window).Sentry;
