+++ resources/js/unpoly/unpoly.ts
@@ -1,0 +1,4 @@
+import "unpoly/unpoly.css";
+import "unpoly/unpoly.js";
+
+export default (<any>window).up;
