+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "experimentalDecorators": true,
+    "module": "commonjs",
+    "target": "es2020",
+    "useDefineForClassFields": false
+  }
+}
